+++ src/shared/components.hpp
@@ -13,6 +13,15 @@ struct ChowKnob : app::SvgKnob {
     }
 };
 
+struct ChowKnobLarge : app::SvgKnob {
+    ChowKnobLarge() {
+        minAngle = -0.76 * M_PI;
+        maxAngle = 0.76 * M_PI;
+        shadow->opacity = 0;
+        setSvg(APP->window->loadSvg(asset::plugin(pluginInstance, "res/components/KnobLarge.svg")));
+    }
+};
+
 struct ChowSmallKnob : app::SvgKnob {
     ChowSmallKnob() {
         minAngle = -0.76 * M_PI;
@@ -42,4 +51,26 @@ struct PinwheelRust : app::SvgSwitch {
 		addFrame(APP->window->loadSvg(asset::plugin(pluginInstance, "res/components/PinwheelRust_0.svg")));
 		addFrame(APP->window->loadSvg(asset::plugin(pluginInstance, "res/components/PinwheelRust_1.svg")));
 	}
+};
+
+struct ChowLabel : ui::Label {
+    std::shared_ptr<Font> font;
+
+    void draw(const DrawArgs& args) override {
+        nvgFontSize(args.vg, fontSize);
+        nvgFontFaceId(args.vg, font->handle);
+        nvgTextLetterSpacing(args.vg, 0.2f);
+        Vec textPos = Vec(0.0f, 0.0f);
+
+        int align = NVG_ALIGN_CENTER;
+        switch(alignment) {
+            case Label::LEFT_ALIGNMENT: align = NVG_ALIGN_LEFT; break;
+            case Label::CENTER_ALIGNMENT: align = NVG_ALIGN_CENTER; break;
+            case Label::RIGHT_ALIGNMENT: align = NVG_ALIGN_RIGHT; break;
+        }
+
+        nvgFillColor(args.vg, color);
+        nvgTextAlign(args.vg, align);
+        nvgText(args.vg, textPos.x, textPos.y, text.c_str(), nullptr);
+    }
 };