+++ CHANGELOG.md
@@ -5,6 +5,7 @@ this file.
 ## [Unreleased]
 - Added new distortion module (Warp)
 - Added variable oversampling options for ChowTape, ChowDer, Werner
+- Fixed control voltage for freq. parameter in ChowModal, now tuned to V/Oct
 - Updated documentation with technical diagrams
 
 ## [1.3.1] - 2020-10-06